// packages/web/src/api/model/jobOverview.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * @6oclock/api
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 1.0.0
 */
import type { JobOverviewOutcome } from './jobOverviewOutcome';
import type { JobOverviewRecentItem } from './jobOverviewRecentItem';

export interface JobOverview {
  id: string;
  description: string;
  queue: string;
  pattern: string;
  enabled: boolean;
  outcome?: JobOverviewOutcome;
  ranAt?: Date;
  ranFor?: number;
  failedReason?: string;
  succeededAt?: Date;
  recent: JobOverviewRecentItem[];
  errors: number;
}